create roles, permissions and policies for website users

+++ app/Http/Controllers/backend/PostController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\Post;
+use App\Models\Topic;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class PostController extends Controller
@@ -27,7 +29,8 @@ class PostController extends Controller
      */
     public function create()
     {
-        return view('backend.post.create');
+        return view('backend.post.create')
+                    ->with('topics', Topic::all());
     }
 
     /**
@@ -38,13 +41,20 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+
         $request->validate([
             'title' => 'required | max:50',
             'sub_title' => 'required | max:50',
-            'description' => 'required'
+            'description' => 'required',
+            'topic' => 'required | array',
         ]);
 
-        Post::create(['title'=>$request->title, 'sub_title'=>$request->sub_title, 'description'=>strip_tags($request->description), 'slug'=>Str()->slug($request->title), 'lang'=>app()->getLocale()]);
+        $post = Post::create(['title'=>$request->title, 'sub_title'=>$request->sub_title, 'description'=>strip_tags($request->description), 'slug'=>Str()->slug($request->title), 'lang'=>app()->getLocale(), 'profile_id'=>Auth::user()->profile->id]);
+        $post->topics()->attach($request->topic);
+
+        foreach($request->image as $image){
+            $post->images()->create(['image'=>$image]);
+        }
 
         return redirect()->route('post.index')->with('success', 'Createed Successfully!');
     }
@@ -69,7 +79,8 @@ class PostController extends Controller
     public function edit(Post $post)
     {
         return view('backend.post.edit')
-                    ->with('post', $post);
+                    ->with('post', $post)
+                    ->with('topics', Topic::all());
     }
 
     /**
@@ -81,17 +92,20 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->authorize('update', $post);
         $request->validate([
             'title' => 'required | max:50',
             'sub_title' => 'required | max:50',
             'description' => 'required',
+            'topic' => 'required | array',
         ]);
 
         $post->title = $request->title;
         $post->sub_title = $request->sub_title;
         $post->description = $request->description;
-
         $post->save();
+
+        $post->topics()->sync($request->topic);
 
         return redirect()->route('post.index')->with('success', 'Updated Successfully!');
     }
@@ -104,11 +118,14 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $post->delete();
         return "success";
     }
 
     public function trash() {
+        $this->authorize('forceDelete', Post::class);
         return view('backend.post.trash')
                     ->with('posts', Post::onlyTrashed()->paginate(10));
     }
@@ -120,6 +137,7 @@ class PostController extends Controller
     }
 
     public function restore($id) {
+        $this->authorize('restore');
         $post = Post::withTrashed()->where('id', $id)->first();
         $post->restore();
         return redirect()->route('post.index');
